refactor(spec): split scale-helpers and tier-2 validate tests
Extract position vs color/fill authoring helpers behind a stable
scale-helpers facade, and split the long validate-tier2 suite into
structure, data, ordering, and M2 files co-located with their concerns.
