+++ .changeset/spec-scale-helpers-tier2-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/spec": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+Split scale authoring helpers into position and color modules with a thin facade, and split the long tier-2 validate suite by concern. Public scale-helper import paths and validation behavior are unchanged.
