+++ packages/spec/tests/validate-tier2-data.test.ts
@@ -1,8 +1,6 @@
 /**
- * Tier-2 validation: structural grammar checks (rule forms, bar y ban,
- * required channels) + data-aware checks (field existence with did-you-mean,
- * all-null columns, stat columns, scale/type compatibility) against inline
- * data or a DataProfile, under documented input limits.
+ * Tier-2 data-aware checks: inline data, DataProfile, input limits, temporal
+ * decision reuse. Production: validate-data*.ts.
  */
 import { fromAny } from "@total-typescript/shoehorn";
 import { describe, expect, it } from "bun:test";
@@ -25,85 +23,6 @@ function errorsOf(input: unknown, options?: Parameters<typeof validate>[1]) {
 function codesOf(input: unknown, options?: Parameters<typeof validate>[1]) {
   return errorsOf(input, options).map((e) => e.code);
 }
-
-describe("tier 2 — structural grammar checks (opt-in, data-free)", () => {
-  it("tier 1 alone accepts a channel-less point layer; tier 2 rejects it", () => {
-    const spec = { layers: [{ geom: "point" }] };
-    expect(validate(spec).ok).toBe(true);
-    expect(codesOf(spec)).toEqual(["missing-required-channel", "missing-required-channel"]);
-  });
-
-  it("text requires x, y, and label", () => {
-    const spec = {
-      aes: { x: { field: "city" }, y: { field: "temp" } },
-      layers: [{ geom: "text" }],
-    };
-    const errors = errorsOf(spec);
-    expect(errors).toHaveLength(1);
-    expect(errors[0]?.code).toBe("missing-required-channel");
-    expect(errors[0]?.path).toBe("/layers/0/aes/label");
-  });
-
-  it("bar with mapped y is rejected with a col suggestion", () => {
-    const errors = errorsOf({
-      aes: { x: { field: "city" }, y: { field: "temp" } },
-      layers: [{ geom: "bar" }],
-    });
-    expect(errors[0]?.code).toBe("computed-y-mapped");
-    expect(errors[0]?.fix?.example).toEqual({ geom: "col" });
-  });
-
-  it("bar with the count-stat y mapping is fine", () => {
-    const result = validate(
-      { aes: { x: { field: "city" }, y: { stat: "count" } }, layers: [{ geom: "bar" }] },
-      {},
-    );
-    expect(result.ok).toBe(true);
-  });
-
-  it("rule: annotation form + a layer-mapped aes.x is ambiguous", () => {
-    expect(
-      codesOf({
-        layers: [{ geom: "rule", aes: { x: { field: "temp" } }, params: { yintercept: 0 } }],
-      }),
-    ).toEqual(["rule-form-ambiguous"]);
-  });
-
-  it("rule: annotation form ignores INHERITED plot aes (inherit.aes = FALSE)", () => {
-    const result = validate(
-      {
-        aes: { x: { field: "temp" }, y: { field: "temp" } },
-        layers: [{ geom: "rule", params: { yintercept: 0 } }],
-      },
-      {},
-    );
-    expect(result.ok).toBe(true);
-  });
-
-  it("rule: neither form is an error", () => {
-    expect(codesOf({ layers: [{ geom: "rule" }] })).toEqual(["rule-form-missing"]);
-  });
-
-  it("rule: both axes mapped is an error", () => {
-    expect(
-      codesOf({
-        aes: { x: { field: "temp" }, y: { field: "temp" } },
-        layers: [{ geom: "rule" }],
-      }),
-    ).toEqual(["rule-both-axes"]);
-  });
-
-  it("rule: null-unset makes the annotation form valid under a mapped plot aes", () => {
-    const result = validate(
-      {
-        aes: { x: { field: "temp" }, y: { field: "temp" } },
-        layers: [{ geom: "rule", aes: { x: null, y: null }, params: { yintercept: 0 } }],
-      },
-      {},
-    );
-    expect(result.ok).toBe(true);
-  });
-});
 
 describe("tier 2 — data-aware checks (inline data)", () => {
   const base = {
@@ -293,166 +212,6 @@ describe("tier 2 — input limits", () => {
   });
 });
 
-describe("tier 2 — cross-stage error ordering (characterization)", () => {
-  // Locks layer-structural → facet → data order so extract refactors cannot
-  // reorder pipeline stages without a deliberate test update.
-  it("reports structural, then facet, then data errors in that order", () => {
-    const errors = errorsOf({
-      data: { values: [{ x: 1, y: 2 }] },
-      layers: [{ geom: "point", aes: { x: { field: "xx" } } }],
-      facet: { wrap: { field: "g" }, rows: { field: "h" } },
-    });
-    expect(errors.map((e) => ({ code: e.code, path: e.path }))).toEqual([
-      { code: "missing-required-channel", path: "/layers/0/aes/y" },
-      { code: "facet-form-ambiguous", path: "/facet" },
-      { code: "unknown-field", path: "/layers/0/aes/x" },
-    ]);
-  });
-
-  it("truncation sentinel stays last when the cap crosses stage boundaries", () => {
-    const errors = errorsOf(
-      {
-        data: { values: [{ x: 1, y: 2 }] },
-        layers: [{ geom: "point", aes: { x: { field: "xx" } } }],
-        facet: { wrap: { field: "g" }, rows: { field: "h" } },
-      },
-      { limits: { maxDiagnostics: 2 } },
-    );
-    expect(errors.map((e) => e.code)).toEqual([
-      "missing-required-channel",
-      "facet-form-ambiguous",
-      "validation-limit",
-    ]);
-    expect(errors.at(-1)?.path).toBe("");
-  });
-
-  it("evidence failure still reports structural errors first, then invalid-data-profile", () => {
-    // Structural stages run before resolveFieldEvidence; a bad profile must
-    // not erase earlier grammar diagnostics or reorder them after data errors.
-    const errors = errorsOf(
-      {
-        layers: [{ geom: "point", aes: { x: { field: "xx" } } }],
-        facet: { wrap: { field: "g" }, rows: { field: "h" } },
-      },
-      { profile: fromAny({ fields: [{ name: "a", type: "numeric" }] }) },
-    );
-    expect(errors.map((e) => e.code)).toEqual([
-      "missing-required-channel",
-      "facet-form-ambiguous",
-      "invalid-data-profile",
-    ]);
-  });
-});
-
-describe("tier 2 — M2 statistical layer (snapshot-tested messages)", () => {
-  it("histogram y mapped to a field -> computed-y-mapped", () => {
-    const errors = errorsOf({
-      layers: [{ geom: "histogram", aes: { x: { field: "v" }, y: { field: "n" } } }],
-    });
-    expect(errors[0]!.code).toBe("computed-y-mapped");
-    expect(errors[0]).toMatchSnapshot();
-  });
-
-  it("density y mapped to a field -> computed-y-mapped", () => {
-    const errors = errorsOf({
-      layers: [{ geom: "density", aes: { x: { field: "v" }, y: { field: "n" } } }],
-    });
-    expect(errors[0]!.code).toBe("computed-y-mapped");
-    expect(errors[0]).toMatchSnapshot();
-  });
-
-  it("bin center + boundary -> bin-center-and-boundary", () => {
-    const errors = errorsOf({
-      layers: [
-        {
-          geom: "histogram",
-          aes: { x: { field: "v" } },
-          params: { binwidth: 1, center: 0, boundary: 0 },
-        },
-      ],
-    });
-    expect(errors[0]!.code).toBe("bin-center-and-boundary");
-    expect(errors[0]).toMatchSnapshot();
-  });
-
-  it("errorbar identity without ymin/ymax -> missing-required-channel (both)", () => {
-    const codes = codesOf({
-      layers: [{ geom: "errorbar", aes: { x: { field: "g" } } }],
-    });
-    expect(codes).toEqual(["missing-required-channel", "missing-required-channel"]);
-  });
-
-  it("errorbar summary requires y instead", () => {
-    const spec = {
-      layers: [{ geom: "errorbar", stat: "summary", aes: { x: { field: "g" } } }],
-    };
-    const errors = errorsOf(spec);
-    expect(errors[0]!.code).toBe("missing-required-channel");
-    expect(errors[0]!.path).toBe("/layers/0/aes/y");
-    expect(errors[0]).toMatchSnapshot();
-  });
-
-  it("smooth and boxplot require x and y", () => {
-    expect(codesOf({ layers: [{ geom: "smooth" }] })).toEqual([
-      "missing-required-channel",
-      "missing-required-channel",
-    ]);
-    expect(codesOf({ layers: [{ geom: "boxplot" }] })).toEqual([
-      "missing-required-channel",
-      "missing-required-channel",
-    ]);
-  });
-
-  it("smooth over nominal fields -> channel-type-mismatch (inline data)", () => {
-    const errors = errorsOf({
-      data: { columns: { c: ["a", "b"], v: [1, 2] } },
-      layers: [{ geom: "smooth", aes: { x: { field: "c" }, y: { field: "v" } } }],
-    });
-    expect(errors[0]!.code).toBe("channel-type-mismatch");
-    expect(errors[0]).toMatchSnapshot();
-  });
-
-  it("boxplot needs discrete x + quantitative y (inline data)", () => {
-    const errors = errorsOf({
-      data: { columns: { c: ["a", "b"], v: [1, 2] } },
-      layers: [{ geom: "boxplot", aes: { x: { field: "v" }, y: { field: "c" } } }],
-    });
-    expect(errors.map((e) => e.code)).toEqual(["channel-type-mismatch", "channel-type-mismatch"]);
-    expect(errors[0]).toMatchSnapshot();
-  });
-
-  it("bin over a nominal x -> channel-type-mismatch (DataProfile)", () => {
-    const profile: DataProfile = { fields: [{ name: "c", type: "nominal" }] };
-    const errors = errorsOf(
-      { layers: [{ geom: "histogram", aes: { x: { field: "c" } } }] },
-      { profile },
-    );
-    expect(errors[0]!.code).toBe("channel-type-mismatch");
-  });
-
-  it("{ stat } channels resolve against the grown STAT_COLUMNS contract", () => {
-    const good = validate(
-      {
-        data: { columns: { v: [1, 2, 3] } },
-        layers: [{ geom: "histogram", aes: { x: { field: "v" }, y: { stat: "ndensity" } } }],
-      },
-      {},
-    );
-    expect(good.ok).toBe(true);
-    const bad = errorsOf({
-      data: { columns: { v: [1, 2, 3] } },
-      layers: [{ geom: "histogram", aes: { x: { field: "v" }, y: { stat: "widthh" } } }],
-    });
-    expect(bad[0]!.code).toBe("unknown-stat-column");
-    expect(bad[0]!.allowed).toEqual(["count", "density", "ncount", "ndensity"]);
-  });
-});
-
-/**
- * Characterization for dataChecks temporal decisions: multi-consumer field
- * reuse, auto+default evidence reuse (ambiguous/invalid/Date), and profile
- * paths that must NOT reuse evidence temporal (always null) when options differ.
- */
 describe("tier 2 — temporal decision reuse (characterization)", () => {
   it("accepts multi-layer charts that share one temporal field under auto time scales", () => {
     const when = ["2024-01-01", "2024-01-02", "2024-01-03"];
